test(server): SystemTagsContract test 追加 + CHANGELOG 加筆 (PR #192 review)
PR #192 の AI レビュー (Cursor) で出た medium 2 件と low 1 件を反映。

medium #2 (テスト不在):
- kSystemTags を mapoi_server/include/mapoi_server/system_tags.hpp に分離
  して header-only 化 (anonymous namespace 廃止)、test から include 可能に
- test_mapoi_server_unit.cpp に SystemTagsContract test を 2 件追加:
  - ExpectedNamesAndOrder: 件数 / 順序 / name の固定検証
  - DescriptionsAreNonEmpty: description が non-empty
  get_tag_definitions service の応答契約を不変条件として固定。受信側
  (mapoi_webui の編集 UI、mapoi_rviz_plugins、custom_tags 排他検査) が
  前提とする 3 system tags の存在を回帰検知できる。

medium #1 (yaml 欠損時挙動差):
- 旧実装: yaml 欠落時 RCLCPP_WARN + 空 list で継続 (= get_tag_definitions
  が 0 system tags を返す degenerate state)
- 新実装: 常に 3 件の compiled-in tag を返す (degenerate state を撤廃)
- CHANGELOG に挙動差として明示。

low #4 (install layout):
- share/mapoi_server/maps/ ディレクトリが install layout から消える点を
  CHANGELOG に追記 (外部スクリプトが share/mapoi_server/maps/ を listing
  していた場合の影響を明示)。

low #3 (mapoi_server_pkg_ 他用途) / low #5 (ログ情報量) は反映なし
(前者は grep 済で他用途なし、後者は許容範囲との指摘)。

Jazzy clean build + test OK (47 tests, 0 errors, 0 failures)。
